Continuing to work through Jack's suggestions.
